requirements.txt
aqtools==0.3.0
aq-snowbyte
boto3==1.37.27
botocore==1.37.27
click==8.1.8
jmespath==1.0.1
numpy==1.26.4
opencv-python==4.11.0.86
pandas==1.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
s3path==0.5.8
s3transfer==0.11.4
six==1.17.0
smart-open==7.1.0
urllib3==2.3.0
wrapt==1.17.2
